% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{ODTP4TMU-package}
\alias{ODTP4TMU}
\alias{ODTP4TMU-package}
\title{ODTP4TMU: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
